+++ mango3-core/src/models/website/mod.rs
@@ -93,11 +93,12 @@ impl Website {
 }
 
 impl Validator {
-    async fn validate_name(&mut self, core_context: &CoreContext, id: Option<Uuid>, value: &str) -> bool {
+    async fn validate_name(&mut self, core_context: &CoreContext, website: Option<&Website>, value: &str) -> bool {
         if self.validate_presence(Input::Name, value)
             && self.validate_length(Input::Name, value, Some(3), Some(255))
             && self.custom_validation(Input::Name, InputError::IsInvalid, &|| Uuid::try_parse(value).is_err())
         {
+            let id = website.map(|w| w.id);
             let name_exists = query!(
                 "SELECT id FROM websites WHERE ($1::uuid IS NULL OR id != $1) AND LOWER(name) = $2 LIMIT 1",
                 id,                   // $1