+++ mango3-leptos-utils/src/components/post_card.rs
@@ -1,6 +1,5 @@
 use leptos::prelude::*;
 
-use crate::i18n::{t, use_i18n};
 use crate::models::PostResp;
 
 #[component]
@@ -9,7 +8,6 @@ pub fn PostCard(
     #[prop(into, default = false)] show_content: bool,
     #[prop(into, optional)] actions: ViewFn,
 ) -> impl IntoView {
-    let i18n = use_i18n();
     let inner_html = move || {
         if show_content {
             post.content_html.clone()
@@ -34,19 +32,15 @@ pub fn PostCard(
                         })
                 }
             } <div class="card-body">
-                <h3 class="card-title">{post.title}</h3>
+                <h3 class="card-title">
+                    <a href=move || {
+                        if !show_content && post.is_published { Some(post.url.clone()) } else { None }
+                    }>{post.title}</a>
+                </h3>
 
                 <div class="prose" inner_html=inner_html />
 
-                <div class="card-actions justify-end">
-                    <Show when=move || !show_content && post.is_published>
-                        <a class="btn btn-ghost font-bold" href=post.url.clone()>
-                            {t!(i18n, shared.view_post)}
-                        </a>
-                    </Show>
-
-                    {actions.run()}
-                </div>
+                <div class="card-actions justify-end">{actions.run()}</div>
             </div>
         </div>
     }