+++ mango3-leptos-utils/src/components/textarea_field.rs
@@ -1,5 +1,6 @@
 use leptos::prelude::*;
 use leptos::text_prop::TextProp;
+use leptos_use::{use_textarea_autosize_with_options, UseTextareaAutosizeOptions, UseTextareaAutosizeReturn};
 
 #[component]
 pub fn TextareaField(
@@ -10,6 +11,17 @@ pub fn TextareaField(
     #[prop(optional, into)] value: Signal<String>,
     name: &'static str,
 ) -> impl IntoView {
+    let node_ref = NodeRef::new();
+
+    let UseTextareaAutosizeReturn {
+        content, set_content, ..
+    } = use_textarea_autosize_with_options(
+        node_ref,
+        UseTextareaAutosizeOptions::default()
+            .style_prop("min-height")
+            .content(value),
+    );
+
     let field_id = move || {
         if let Some(id) = id {
             id.to_owned()
@@ -26,12 +38,14 @@ pub fn TextareaField(
                 <span class="label-text">{move || label.get()}</span>
             </label>
             <textarea
+                node_ref=node_ref
+                prop:value=content
+                on:input=move |event| set_content.set(event_target_value(&event))
                 class="textarea textarea-bordered"
                 class:textarea-error=has_error
                 id=field_id
                 name=name
                 rows=rows
-                prop:value=value
             />
             <div class="label">
                 <span class="label-text-alt text-error">{move || error.get()}</span>