+++ mango3-leptos-utils/src/pages/not_found_page.rs
@@ -6,9 +6,9 @@ use super::Page;
 pub fn NotFoundPage() -> impl IntoView {
     view! {
         <Page title="Error 404: Page not found">
-            <h2>"Whoops!"</h2>
+            <h2 class="h2">"Whoops!"</h2>
 
-            <h3>"There is nothing here!"</h3>
+            <h3 class="h3">"There is nothing here!"</h3>
         </Page>
     }
 }