+++ mango3-studio/src/constants.rs
@@ -1,1 +1,2 @@
+pub const KEY_PARAM_POST_ID: &str = "post_id";
 pub const KEY_PARAM_WEBSITE_ID: &str = "website_id";