+++ mango3-studio/src/pages/websites/mod.rs
@@ -1,10 +1,12 @@
 mod edit_page;
+mod edit_post_page;
 mod new_post_page;
 mod posts_page;
 mod show_page;
 mod show_parent_page;
 
 pub use edit_page::EditPage;
+pub use edit_post_page::EditPostPage;
 pub use new_post_page::NewPostPage;
 pub use posts_page::PostsPage;
 pub use show_page::ShowPage;