+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - Features to create and show posts.
 - Feature to reset password.
 - Feature to register only with invitation code.
+- Feature to update and delete posts.